Remove/Leave Conversation Implemented
-removes the convo if there were only 2 people 
-if GC, removes the user from the current group leaving the rest still in the GC

+++ lib/database/flutterfire.dart
@@ -90,12 +90,24 @@ Future<String> getName(String userId) async {
   }
 }
 
-Future<bool> deleteConversation(String conversationId) async {
+Future<bool> leaveConversation(String conversationId) async {
   try {
-    await FirebaseFirestore.instance.collection('Conversations').doc(conversationId).delete();
+    // await FirebaseFirestore.instance.collection('Conversations').doc(conversationId).delete();
+    DocumentSnapshot convoSnapshot = await FirebaseFirestore.instance.collection('Conversations').doc(conversationId).get();
+    List<dynamic> people = convoSnapshot.data()['people'];
+
+    if (people.length <= 2) // If only two people are in the convo
+      await FirebaseFirestore.instance.collection('Conversations').doc(conversationId).delete();
+    else {
+      // If Group chat
+      DocumentReference convoRef = FirebaseFirestore.instance.collection('Conversations').doc(conversationId);
+      convoRef.update({
+        'people': FieldValue.arrayRemove([FirebaseAuth.instance.currentUser.uid])
+      });
+    }
     return true;
   } on Exception catch (e) {
-    print("Deletion failed! Error: $e");
+    print("Deletion/Leave failed! Error: $e");
   }
   return false;
 }
